WIP src/digest: CRC and similar digest routines
This contains both the standard algorithms (such as CRC-16-CCITT) and
special non-standard algorithms designed to be as small (in code size)
as possible.

XXX It also contains various wrappers to generate digests from memory.

It's not clear if in digest.a65 forcing the listing off (and adding
{EXPAND} to all macros) is the best thing to do, since it's
conceivable that a caller might want all the macros listed anyway. But
that doesn't seem too likely, given that most callers will probably
use macros only once each, and use only some of them.